requirements.txt
flask==3.1.1
PyMuPDF==1.25.5
ebooklib==0.20
Pillow>=10.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-docx>=1.1.2
pdfplumber>=0.11.0
wordfreq>=3.1.1
pyphen>=0.17.2
rfc3986>=2.0.0
tldextract>=5.1.2
